requirements.txt
streamlit>=1.18.0
openai==0.28.0
pandas
matplotlib
python-dotenv